Chore: Migrate app-cli cli-utils to TypeScript
The 14+ consumers all destructure via require() so the named
export shape (`export const cliUtils: CliUtils = ...`) is
preserved without altering any call site. Defined CliUtils as
an interface so the method-assignments at file scope are
type-checked against a single contract.

Several minor type-driven adjustments:

- The dead `if (i >= a.length) { output[a.name] = null }` branch
  in makeCommandArgs is removed and its always-taken else branch
  is kept inline. `a` was {required, name} (no length property),
  so the if was unreachable. See review-later.md for the likely-
  intended fix (probably meant `args['_'].length`).
- `printArray`'s `if (!rows.length) return '';` early exit is
  redundant — the inner loops naturally skip when rows is empty.
  Removed to make the function cleanly void.
- `stdoutLogger`'s `addTarget('console', ...)` is updated to
  `TargetType.Console`; the partial-Console object passed in is
  cast through unknown to Console (same pattern as command-server).
- `prompt()`'s unused `initialText` param is underscore-prefixed
  so tsc's noUnusedParameters accepts it. The original eslint-
  disable comment doesn't suppress tsc-side errors.
- The Writable subclass synthesised in `prompt()` carries a
  runtime-added `muted` flag that the base type doesn't know
  about; `this` inside the write callback is cast to any with a
  scoped eslint-disable.

cliUtils.parseFlags/parseCommandArg now have explicit return
types (ParsedFlags, ParsedCommandArg) that consumers can import
if needed in the future.

Change by Claude Code
