+++ .eslintignore
@@ -101,6 +101,7 @@ packages/app-cli/app/app.js
 packages/app-cli/app/autocompletion.js
 packages/app-cli/app/base-command.js
 packages/app-cli/app/cli-integration-tests.test.js
+packages/app-cli/app/cli-utils.js
 packages/app-cli/app/command-apidoc.js
 packages/app-cli/app/command-attach.js
 packages/app-cli/app/command-batch.js
